Deleted event file moved code to state file
