+++ lib/auth/confirm/confirmation_bloc.dart
@@ -7,7 +7,6 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import '../auth_cubit.dart';
 import '../auth_repository.dart';
 import '../form_submission_status.dart';
-import 'confirmation_event.dart';
 import 'confirmation_state.dart';
