+++ lib/auth/login/login_state.dart
@@ -43,3 +43,29 @@ class LoginState {
     );
   }
 }
+
+abstract class LoginEvent {}
+
+/// Event when user changes the text in the username field
+class LoginUsernameChanged extends LoginEvent {
+  final String username;
+
+  LoginUsernameChanged({this.username});
+}
+
+/// Event when user changes the text in the password field
+class LoginPasswordChanged extends LoginEvent {
+  final String password;
+
+  LoginPasswordChanged({this.password});
+}
+
+/// Event when standard login button is pressed
+class LoginSubmitted extends LoginEvent {}
+
+/// Event when login with facebook button is pressed
+class LoginFacebook extends LoginEvent {}
+
+/// Event when login with google button is pressed
+/// not currently used
+class LoginGoogle extends LoginEvent {}
