+++ lib/auth/login/login_view.dart
@@ -15,7 +15,6 @@ import '../auth_cubit.dart';
 import '../auth_repository.dart';
 import '../form_submission_status.dart';
 import 'login_bloc.dart';
-import 'login_event.dart';
 import 'login_state.dart';
 
 /// Holds the login view widget tree
