+++ lib/auth/sign_up/sign_up_state.dart
@@ -7,6 +7,36 @@
 
 import '../form_submission_status.dart';
 
+
+
+abstract class SignUpEvent {}
+
+/// Event when user changes the text in the username field
+class SignUpUsernameChanged extends SignUpEvent {
+  final String username;
+
+  SignUpUsernameChanged({this.username});
+}
+
+/// Event when user changes the text in the email field
+class SignUpEmailChanged extends SignUpEvent {
+  final String email;
+
+  SignUpEmailChanged({this.email});
+}
+
+/// Event when user changes the text in the password field
+class SignUpPasswordChanged extends SignUpEvent {
+  final String password;
+
+  SignUpPasswordChanged({this.password});
+}
+
+/// Event when the sign up button is pressed
+class SignUpSubmitted extends SignUpEvent {}
+
+
+
 /// Holds the current state of the sign up form
 class SignUpState {
   final String username;
@@ -49,3 +79,5 @@ class SignUpState {
     );
   }
 }
+
+
